# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AMX ERP Suite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory/products:
    get:
      operationId: listProducts
      tags: [inventory]
      summary: List products
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [inventory]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /inventory/products/{id}:
    get:
      operationId: getProduct
      tags: [inventory]
      summary: Get product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [inventory]
      summary: Update product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [inventory]
      summary: Delete product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{projectId}/tasks:
    get:
      operationId: listTasks
      tags: [projects]
      summary: List tasks for project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [projects]
      summary: Create task
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /projects/{projectId}/tasks/{taskId}:
    patch:
      operationId: updateTask
      tags: [projects]
      summary: Update task
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
        - in: path
          name: taskId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [projects]
      summary: Delete task
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
        - in: path
          name: taskId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Analytics
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, sku, category, quantity, price, reorderLevel, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: string
        category:
          type: string
        quantity:
          type: integer
        price:
          type: number
        reorderLevel:
          type: integer
        status:
          type: string
        createdAt:
          type: string
        description:
          type: ["string", "null"]
    ProductInput:
      type: object
      required: [name, sku, category, quantity, price, reorderLevel, status]
      properties:
        name:
          type: string
        sku:
          type: string
        category:
          type: string
        quantity:
          type: integer
        price:
          type: number
        reorderLevel:
          type: integer
        status:
          type: string
        description:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        quantity:
          type: integer
        price:
          type: number
        reorderLevel:
          type: integer
        status:
          type: string
        description:
          type: string
...
    Project:
      type: object
      required: [id, name, description, status, priority, startDate, endDate, budget, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        budget:
          type: number
        createdAt:
          type: string
        progress:
          type: integer
    ProjectInput:
      type: object
      required: [name, description, status, priority, startDate, endDate, budget]
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        budget:
          type: number
        progress:
          type: integer
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        endDate:
          type: string
        budget:
          type: number
        progress:
          type: integer
    Task:
      type: object
      required: [id, projectId, title, status, priority, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        assignee:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
    TaskInput:
      type: object
      required: [title, status, priority]
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        assignee:
          type: string
        dueDate:
          type: string
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
        priority:
          type: string
        assignee:
          type: string
        dueDate:
          type: string

    # Analytics
...
    Notification:
      type: object
      required: [id, title, message, type, read, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
        read:
          type: boolean
        createdAt:
          type: string

    # Forecast
...