# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AMX ERP Suite API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  sku: string;
  category: string;
  quantity: number;
  price: number;
  reorderLevel: number;
  status: string;
  createdAt: string;
  /** @nullable */
  description?: string | null;
}